Added class Isoteck_micro.cs ...removed class F26
Added a new class for the ISOtech bridge and removed the old f26 from the program as this is being decommisioned.  Most of the changes to code are in the Isoteck_micro.cs class.  This bridges uses a different command structure to the old F26 bridge.  The bridge also has and itegral 10:1 multiplexor which is access on channels 10 through 19.  The bridge is set to use the internal resistor calibrated by NPL.  The bridge operates in ratio mode.  I have checked that the ratio of the Tinsley resistor to the internal resistor is correct; it is.